More controlled ID==self check
