+++ service/bootstrap_master.go
@@ -41,7 +41,7 @@ func (s *Service) bootstrapMaster(ctx context.Context, runner Runner, config Con
 	if err != nil {
 		s.log.Fatalf("Cannot find HTTP server info: %#v", err)
 	}
-	if !IsPortOpen(containerHTTPPort) {
+	if !WaitUntilPortAvailable(containerHTTPPort, time.Second*5) {
 		s.log.Fatalf("Port %d is already in use", containerHTTPPort)
 	}
