+++ service/bootstrap_slave.go
@@ -104,7 +104,7 @@ func (s *Service) bootstrapSlave(peerAddress string, runner Runner, config Confi
 	if err != nil {
 		s.log.Fatalf("Cannot find HTTP server info: %#v", err)
 	}
-	if !IsPortOpen(containerHTTPPort) {
+	if !WaitUntilPortAvailable(containerHTTPPort, time.Second*5) {
 		s.log.Fatalf("Port %d is already in use", containerHTTPPort)
 	}
