+++ service/port_check.go
@@ -25,6 +25,7 @@ package service
 import (
 	"fmt"
 	"net"
+	"time"
 )
 
 // IsPortOpen checks if a TCP port is free to listen on.
@@ -35,4 +36,20 @@ func IsPortOpen(port int) bool {
 	}
 	l.Close()
 	return true
+}
+
+// WaitUntilPortAvailable waits until a TCP port is free to listen on
+// or a timeout occurs.
+// Returns true when port is free to listen on.
+func WaitUntilPortAvailable(port int, timeout time.Duration) bool {
+	//start := time.Now()
+	for {
+		if IsPortOpen(port) {
+			return true
+		}
+		//if time.Since(start) >= timeout {
+		return false
+		//}
+		//time.Sleep(time.Millisecond * 10)
+	}
 }
